requirements.txt
Django==1.8
Fabric==1.10.2
Pillow==2.9.0
Pygments==2.0.2
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-compressor==1.4
django-debug-toolbar==1.3.0
django-grappelli==2.7.1
djangorestframework==3.0.0
djrill==1.4.0
drf-nested-routers==0.9.0
ecdsa==0.13
fabtools==0.19.0
gunicorn==19.1.1
html5lib==0.999999
httpie==0.9.2
loremipsum==1.0.5
paramiko==1.16.0
psycopg2==2.6.1
pycrypto==2.6.1
requests==2.8.1
six==1.8.0
sqlparse==0.1.18
static3==0.5.1
